Refactoring of Anthropic providers
