+++ extensions/positron-assistant/src/providers/base/modelProvider.ts
